pyproject.toml
[project]
name = "muphrid"
version = "0.1.0"
description = "LLM agent for autonomous astrophotography post-processing"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
dependencies = [
    "astropy>=7.2.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.5",
    "langchain-core>=1.2.17",
    "langchain-openai>=1.1.10",
    "langgraph>=1.1.2",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "numpy>=2.4.2",
    "photutils>=2.3.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pyexiftool>=0.5.6",
    "python-dotenv>=1.2.2",
    "pywavelets>=1.9.0",
    "rich>=14.3.3",
    "scikit-image>=0.26.0",
    "gradio>=5.0",
    "typer>=0.24.1",
    "sqlite-vec>=0.1.6",
    "ollama>=0.4.0",
    "fastembed>=0.4.0",
    "aiosqlite>=0.22.1",
]

[project.scripts]
muphrid = "muphrid.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
